+++ OwnMyWay/OwnMyWay/Data/Persistent/CoreData/DataMapping/LocationMO+CoreDataClass.swift
@@ -5,7 +5,6 @@
 //  Created by 김우재 on 2021/11/11.
 //
 //
-
 import CoreData
 import Foundation
