+++ OwnMyWay/OwnMyWay/Data/Persistent/CoreData/DataMapping/RecordMO+CoreDataClass.swift
@@ -5,7 +5,6 @@
 //  Created by 김우재 on 2021/11/15.
 //
 //
-
 import CoreData
 import Foundation
