+++ OwnMyWay/OwnMyWay/Data/Persistent/CoreData/RecordMO+CoreDataProperties.swift
@@ -5,7 +5,6 @@
 //  Created by 김우재 on 2021/11/25.
 //
 //
-
 import Foundation
 import CoreData
